// src/clients/js/types/outboundEvent.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  fixDecoderSize,
  fixEncoderSize,
  getBytesDecoder,
  getBytesEncoder,
  getStructDecoder,
  getStructEncoder,
  getU32Decoder,
  getU32Encoder,
  getU64Decoder,
  getU64Encoder,
  getU8Decoder,
  getU8Encoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  type ReadonlyUint8Array,
} from "@solana/kit";

export type OutboundEvent = {
  discriminator: number;
  networkIn: number;
  networkOut: number;
  tokenIn: ReadonlyUint8Array;
  tokenOut: ReadonlyUint8Array;
  fromAddress: ReadonlyUint8Array;
  toAddress: ReadonlyUint8Array;
  amount: bigint;
  relayerFee: bigint;
  nonce: ReadonlyUint8Array;
};

export type OutboundEventArgs = {
  discriminator: number;
  networkIn: number;
  networkOut: number;
  tokenIn: ReadonlyUint8Array;
  tokenOut: ReadonlyUint8Array;
  fromAddress: ReadonlyUint8Array;
  toAddress: ReadonlyUint8Array;
  amount: number | bigint;
  relayerFee: number | bigint;
  nonce: ReadonlyUint8Array;
};

export function getOutboundEventEncoder(): FixedSizeEncoder<OutboundEventArgs> {
  return getStructEncoder([
    ["discriminator", getU8Encoder()],
    ["networkIn", getU32Encoder()],
    ["networkOut", getU32Encoder()],
    ["tokenIn", fixEncoderSize(getBytesEncoder(), 32)],
    ["tokenOut", fixEncoderSize(getBytesEncoder(), 32)],
    ["fromAddress", fixEncoderSize(getBytesEncoder(), 32)],
    ["toAddress", fixEncoderSize(getBytesEncoder(), 32)],
    ["amount", getU64Encoder()],
    ["relayerFee", getU64Encoder()],
    ["nonce", fixEncoderSize(getBytesEncoder(), 32)],
  ]);
}

export function getOutboundEventDecoder(): FixedSizeDecoder<OutboundEvent> {
  return getStructDecoder([
    ["discriminator", getU8Decoder()],
    ["networkIn", getU32Decoder()],
    ["networkOut", getU32Decoder()],
    ["tokenIn", fixDecoderSize(getBytesDecoder(), 32)],
    ["tokenOut", fixDecoderSize(getBytesDecoder(), 32)],
    ["fromAddress", fixDecoderSize(getBytesDecoder(), 32)],
    ["toAddress", fixDecoderSize(getBytesDecoder(), 32)],
    ["amount", getU64Decoder()],
    ["relayerFee", getU64Decoder()],
    ["nonce", fixDecoderSize(getBytesDecoder(), 32)],
  ]);
}

export function getOutboundEventCodec(): FixedSizeCodec<
  OutboundEventArgs,
  OutboundEvent
> {
  return combineCodec(getOutboundEventEncoder(), getOutboundEventDecoder());
}