// src/clients/js/types/overrideOutboundEvent.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  fixDecoderSize,
  fixEncoderSize,
  getBytesDecoder,
  getBytesEncoder,
  getStructDecoder,
  getStructEncoder,
  getU64Decoder,
  getU64Encoder,
  getU8Decoder,
  getU8Encoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  type ReadonlyUint8Array,
} from "@solana/kit";

export type OverrideOutboundEvent = {
  discriminator: number;
  toAddress: ReadonlyUint8Array;
  relayerFee: bigint;
  nonce: ReadonlyUint8Array;
};

export type OverrideOutboundEventArgs = {
  discriminator: number;
  toAddress: ReadonlyUint8Array;
  relayerFee: number | bigint;
  nonce: ReadonlyUint8Array;
};

export function getOverrideOutboundEventEncoder(): FixedSizeEncoder<OverrideOutboundEventArgs> {
  return getStructEncoder([
    ["discriminator", getU8Encoder()],
    ["toAddress", fixEncoderSize(getBytesEncoder(), 32)],
    ["relayerFee", getU64Encoder()],
    ["nonce", fixEncoderSize(getBytesEncoder(), 32)],
  ]);
}

export function getOverrideOutboundEventDecoder(): FixedSizeDecoder<OverrideOutboundEvent> {
  return getStructDecoder([
    ["discriminator", getU8Decoder()],
    ["toAddress", fixDecoderSize(getBytesDecoder(), 32)],
    ["relayerFee", getU64Decoder()],
    ["nonce", fixDecoderSize(getBytesDecoder(), 32)],
  ]);
}

export function getOverrideOutboundEventCodec(): FixedSizeCodec<
  OverrideOutboundEventArgs,
  OverrideOutboundEvent
> {
  return combineCodec(
    getOverrideOutboundEventEncoder(),
    getOverrideOutboundEventDecoder(),
  );
}